Huge update
After the optimization of almost the double of performance on Dadd() and Dsub() the other day (tested with 999999999 samples of each case (4 [999999999 * 4])), I decided to recreate the String_add() and String_sub() functions, now, they seem to work fine.

I tested String_add() for a while and worked visibly fine, I might do an automated testing tool to see if there is any possible case this is not working. But I haven't tested String_sub(), as it should work properly, its just a fork of String_add().

Also, added the respective options to help() and added lost of documentation inside code comments, I'll have to make a .md file to organize it.